Close to finished
It compiles and runs, scanner and parser are pretty close to being
done. Just need to make some refinements. I did what I could figure out
for the tree and special but got a little confused in the process. Just
have to wait for him to cover it in class